third_party/blink/web_tests/external/wpt/trusted-types/Document-write-appending-line-feed-expected.txt

This is a testharness.js-based test.
[FAIL] document.write() with TrustedHTML arguments only.
  assert_equals: expected "(1)(2)(3)(4)" but got "[(1)(2)(3)(4)]"
[FAIL] document.writeln() with TrustedHTML arguments only.
  assert_equals: expected "(1)(2)(3)(4)\\n" but got "[(1)(2)(3)(4)]\\n"
Harness: the test ran to completion.
